+++ src/runtime_src/tools/xclbin/XclBinUtilities.h
@@ -27,6 +27,7 @@
 
 
 #include <iostream>
+#include <stdint.h>
 namespace XclBinUtilities {
 //
 template<typename ... Args>
@@ -58,13 +59,13 @@ bool findBytesInStream(std::fstream& _istream, const std::string& _searchString,
 void setVerbose(bool _bVerbose);
 void TRACE(const std::string& _msg, bool _endl = true);
 void TRACE_PrintTree(const std::string& _msg, const boost::property_tree::ptree& _pt);
-void TRACE_BUF(const std::string& _msg, const char* _pData, unsigned long _size);
+void TRACE_BUF(const std::string& _msg, const char* _pData, uint64_t _size);
 
 void safeStringCopy(char* _destBuffer, const std::string& _source, unsigned int _bufferSize);
-unsigned int bytesToAlign(unsigned int _offset);
+unsigned int bytesToAlign(uint64_t _offset);
 unsigned int alignBytes(std::ostream & _buf, unsigned int _byteBoundary);
 
-void binaryBufferToHexString(const unsigned char* _binBuf, unsigned int _size, std::string& _outputString);
+void binaryBufferToHexString(const unsigned char* _binBuf, uint64_t _size, std::string& _outputString);
 void hexStringToBinaryBuffer(const std::string& _inputString, unsigned char* _destBuf, unsigned int _bufferSize);
 uint64_t stringToUInt64(const std::string& _sInteger);
 void printKinds();
